fix type definitions

+++ SC16IS750.h
@@ -61,7 +61,7 @@ class SC16IS750 {
     // int readPin(int pin);
     // set crystal frequency before setting the baud rate
     void setCyrstalFrequencyMHz(float freqMHz);
-    void configureUart(uint16_t baudRate = 9600);
+    void configureUart(uint32_t baudRate = 9600);
     bool uartConnected();
     void setChannel(uint8_t channel);
